[REFACTOR] #53 : contract camera package로 위치 수정
