+++ client/app/src/main/java/com/example/healthc/presentation/camera/contract/PickSinglePhotoContract.kt
@@ -1,9 +1,11 @@
-package com.example.healthc.presentation.utils
+package com.example.healthc.presentation.camera.contract
 
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.net.Uri
+import android.os.Build
+import android.os.ext.SdkExtensions
 import android.provider.MediaStore
 import androidx.activity.result.contract.ActivityResultContract
 
@@ -25,5 +27,20 @@ class PickSinglePhotoContract : ActivityResultContract<Void?, Uri?>() {
 
     override fun parseResult(resultCode: Int, intent: Intent?): Uri? {
         return intent.takeIf { resultCode == Activity.RESULT_OK }?.data
+    }
+
+    object PhotoPickerAvailabilityChecker {
+
+        private const val ANDROID_R_REQUIRED_EXTENSION_VERSION = 2
+
+        fun isPhotoPickerAvailable(): Boolean {
+            return when {
+                Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU -> true
+                Build.VERSION.SDK_INT >= Build.VERSION_CODES.R -> {
+                    SdkExtensions.getExtensionVersion(Build.VERSION_CODES.R) >= ANDROID_R_REQUIRED_EXTENSION_VERSION
+                }
+                else -> false
+            }
+        }
     }
 }
